requirements.txt
Django==4.0
djangorestframework==3.14.0
drf-yasg==1.21.4
django-cors-headers==3.13.0
psycopg2-binary==2.8.6